backend/requirements.txt
Minimal, pinned requirements for Hermes backend (cleaned)
Adjust versions as needed; use a lockfile for deterministic builds

fastapi==0.95.2
uvicorn[standard]==0.22.0
pydantic==2.11.9
pydantic-settings==2.11.0
python-dotenv==1.1.1
requests==2.32.5
numpy==2.3.3
pandas==2.2.3

# Finance & data
yfinance==0.2.25
alpha-vantage==2.3.0

# OpenAI & agents (note: openai-agents may pin openai to 1.x)
openai==1.109.1
openai-agents==0.3.3

# Visualization
graphviz==0.21

# ML helpers - transformers optional (large)
transformers==4.56.2
joblib==1.3.2

# Web form uploads
python-multipart==0.0.20

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.10

# HTTP async clients
aiohttp==3.9.1
httpx==0.25.2

# Optional (enable only on supported build machines):
# torch==2.2.0
# lightgbm==4.0.0

# Dev/test
pytest>=7.4.0
pytest-asyncio>=0.21.0
